wip(host-v0.3.0) feat(https) DELETE/POST sessions
To start and stop sessions.

Sessions are started just after theme is selected and after each
transcription loop.  At the end of the session the cycle can start over
if the client remembers its theme.  Currently there is no locking system
to disable multiple sessions on the same theme, therefore extra care is
required at this stage of the development.